requirements.txt
Django==3.1.6
pymongo==3.3.0
djangorestframework==3.11.2
stripe==2.55.2
cryptography==3.1.1
urllib3==1.26.3